tuer/requirements.txt
paho-mqtt
tweepy
blinker
homie
pyserial